results

+++ DAL/CupCompetitionTypeRepository.cs
@@ -9,9 +9,11 @@ namespace DAL
     /// <summary>
     /// DAL класс для работы таблицей CupCompetitionType
     /// </summary>
-    public class CupCompetitionTypeRepository
+    public class CupCompetitionTypeRepository : BaseRepository<CupCompetitionTypeParams, CupCompetitionType>
     {
-        private CupCompetitionTypeParams Convert(CupCompetitionType cupCompType)
+        protected override Func<CupCompetitionType, int> GetPrimaryKeyValue => (x) => { return x.Id; };
+
+        protected override CupCompetitionTypeParams ConvertToModel(CupCompetitionType cupCompType)
         {
             return new CupCompetitionTypeParams
             {
@@ -22,7 +24,7 @@ namespace DAL
             };
         }
 
-        private CupCompetitionType Convert(CupCompetitionTypeParams cupCompType)
+        protected override CupCompetitionType ConvertToEntity(CupCompetitionTypeParams cupCompType)
         {
             return new CupCompetitionType
             {
@@ -33,62 +35,6 @@ namespace DAL
             };
         }
 
-        public ResultInfo Add(CupCompetitionTypeParams cupCompType)
-        {
-            var res = new ResultInfo();
-
-            try
-            {
-                using (var db = DBContext.GetContext())
-                {
-                    db.CupCompetitionType.Add(Convert(cupCompType));
-                    db.SaveChanges();
-                }
-            }
-            catch (Exception exc)
-            {
-                res.IsOk = false;
-                res.ErrorMessage = "Произошла ошибка при добавлении записи в таблицу CupCompetitionType";
-                res.Exc = exc;
-            }
-
-            return res;
-        }
-
-        /// <summary>
-        /// Обновить упражнение на соревновании
-        /// </summary>
-        /// <param name="id">ид. упражнения на соревновании</param>
-        /// <param name="cup">соревнование</param>
-        /// <returns></returns>
-        public ResultInfo Update(int id, CupCompetitionTypeParams cup)
-        {
-            var res = new ResultInfo();
-
-            try
-            {
-                using (var db = DBContext.GetContext())
-                {
-                    var updating = new CupCompetitionType { Id = id};
-                    db.CupCompetitionType.Attach(updating);
-
-                    updating.IdCompetitionType = cup.IdCompetitionType;
-                    updating.IdCup = cup.IdCup;
-                    updating.TimeFirstShift = cup.TimeFirstShift;
-
-                    db.Entry(updating).State = EntityState.Modified;
-
-                    db.SaveChanges();
-                }
-            }
-            catch (Exception exc)
-            {
-                throw new Exception("Произошла ошибка при получении списка упражнений на соревновании");
-            }
-
-            return res;
-        }
-
         /// <summary>
         /// Получить список упржанений по соревнованию
         /// </summary>
@@ -96,22 +42,7 @@ namespace DAL
         /// <returns></returns>
         public List<CupCompetitionTypeParams> GetByCup(int idCup)
         {
-            var list = new List<CupCompetitionTypeParams>();
-
-            try
-            {
-                using (var db = DBContext.GetContext())
-                {
-                    var query = db.CupCompetitionType.Where(x => x.IdCup == idCup).ToList();
-                    list = query.ConvertAll(Convert);
-                }
-            }
-            catch (Exception exc)
-            {
-                throw new Exception("Произошла ошибка при получении списка упражнений на соревновании");
-            }
-
-            return list;
+            return GetFiltered(x => x.IdCup == idCup);
         }
 
         /// <summary>
@@ -121,7 +52,7 @@ namespace DAL
         /// <returns></returns>
         public List<CupCompetitionTypeParams> GetByCupWithEntry(int idCup)
         {
-            var res = new List<CupCompetitionTypeParams>();
+            List<CupCompetitionTypeParams> res;
 
             try
             {
@@ -131,7 +62,7 @@ namespace DAL
                                  join cupCompetition in db.CupCompetitionType on entry.IdCupCompetitionType equals cupCompetition.Id
                                  where cupCompetition.IdCup == idCup
                                  select cupCompetition).ToList();
-                    res = query.ConvertAll(Convert);
+                    res = query.Select(ConvertToModel).ToList();
                 }
             }
             catch (Exception exc)
@@ -148,56 +79,9 @@ namespace DAL
         /// <param name="idCup">ид. соревнования</param>
         /// <param name="idCompType">ид. типа</param>
         /// <returns></returns>
-        public CupCompetitionType GetByCupAndCompType(int idCup, int idCompType)
+        public CupCompetitionTypeParams GetByCupAndCompType(int idCup, int idCompType)
         {
-            CupCompetitionType res;
-
-            try
-            {
-                using (var db = DBContext.GetContext())
-                {
-                    res = db.CupCompetitionType.Single(x => x.IdCup == idCup && x.IdCompetitionType == idCompType);
-                }
-            }
-            catch (Exception exc)
-            {
-                throw new Exception("Произошла ошибка при получении CupCompetitionType");
-            }
-
-            return res;
-        }
-
-        /// <summary>
-        /// Добавить список
-        /// </summary>
-        /// <param name="adding">добавляемые</param>
-        /// <returns></returns>
-        public ResultInfo AddRange(List<CupCompetitionTypeParams> adding)
-        {
-            var res = new ResultInfo();
-
-            try
-            {
-                using (var db = DBContext.GetContext())
-                {
-                    var dalAdding = adding.ConvertAll(Convert).ToList();
-
-                    foreach (var item in dalAdding)
-                    {
-                        db.CupCompetitionType.Add(item);
-                    }
-
-                    db.SaveChanges();
-                }
-            }
-            catch (Exception exc)
-            {
-                res.IsOk = false;
-                res.ErrorMessage = "Произошла ошибка при добавлении записи в таблицу CupCompetitionType";
-                res.Exc = exc;
-            }
-
-            return res;
+            return GetFiltered(x => x.IdCup == idCup && x.IdCompetitionType == idCompType).FirstOrDefault();
         }
 
         /// <summary>
@@ -216,8 +100,11 @@ namespace DAL
 
                     foreach (var item in deleting)
                     {
-                        var delItem = db.CupCompetitionType.Where(x => x.Id == item.Id).Single();
-                        db.CupCompetitionType.Remove(delItem);
+                        var delItem = db.CupCompetitionType.FirstOrDefault(x => x.Id == item.Id);
+                        if (delItem != null)
+                        {
+                            db.CupCompetitionType.Remove(delItem);
+                        }
                     }
 
                     db.SaveChanges();
